- update user branch default selected - chg db time from bigInt to datetime - ticket and queue table seeder - chg foreign key from username to id

+++ database/migrations/2018_01_30_165745_create_changes_table.php
@@ -18,7 +18,7 @@ class CreateChangesTable extends Migration
             $table->engine = 'InnoDB';
             $table->bigIncrements('id')->unique();
             $table->String('change');
-            $table->bigInteger('time');
+            $table->integer('time');
             $table->text('ticket_ids')->nullable();
             $table->timestamps();
         });
